requirements.txt
numpy
opencv-python
scikit-image
pillow
fastapi
uvicorn
pydantic
matplotlib
pytest
black
ruff
